New SummaryView with plant capacities
