+++ DistrictEnergy/ViewModels/ResultsViewModel.cs
@@ -5,7 +5,6 @@ using System.Drawing.Printing;
 using System.Globalization;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Windows.Data;
 using System.Windows.Media;
 using Deedle;
 using DistrictEnergy.Helpers;
@@ -829,184 +828,6 @@ namespace DistrictEnergy.ViewModels
         public ChartMode ChartMode { get; set; }
 
         #endregion
-    }
-
-    /// <summary>
-    ///     Converts a kWh quantity to MWh or GWh depending on magnitude of value
-    /// </summary>
-    public class KWhConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999)
-                    return (d / 1000).ToString("N1", culture); // for MWh
-
-                if (d > 999999)
-                    return (d / 1000000).ToString("N1", culture); // for GWh
-                return d.ToString("N1", culture); // for kWh;
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            double d;
-            if (double.TryParse((string) value, out d))
-                return d;
-            return 0.0;
-        }
-    }
-
-    /// <summary>
-    ///     Converts a kWh quantity to MWh or GWh depending on magnitude of value
-    /// </summary>
-    public class kgCO2ConverterUnit : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999)
-                    return "tCO2eq";
-
-                if (d > 999999)
-                    return "MtCO2eq";
-                return "kgCO2eq";
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return "kWh";
-        }
-    }
-
-    /// <summary>
-    ///     Converts a kWh qauntity to MWh or GWh depending on magnitude of value
-    /// </summary>
-    public class NormalizedkgCO2ConverterUnit : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999)
-                    return "tCO2eq/m²";
-
-                if (d > 999999)
-                    return "MtCO2eq";
-                return "kgCO2eq/m²";
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return "kgCO2eq";
-        }
-    }
-
-    /// <summary>
-    ///     Converts a kW quantity to MWh or GWh depending on magnitude of value
-    /// </summary>
-    public class KWConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999)
-                    return (d / 1000).ToString("N1", culture); // for MW
-
-                if (d > 999999)
-                    return (d / 1000000).ToString("N1", culture); // for GW
-                return d.ToString("N1", culture); // for kW;
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            double d;
-            if (double.TryParse((string) value, out d))
-                return d;
-            return 0.0;
-        }
-    }
-
-    /// <summary>
-    ///     Converts a kW qauntity to MW or GW depending on magnitude of value
-    /// </summary>
-    public class KWConverterUnit : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999)
-                    return "MW";
-
-                if (d > 999999)
-                    return "GW";
-                return "kW";
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return "kW";
-        }
-    }
-
-    /// <summary>
-    ///     If value is higher than 999999 m2, than show square km2
-    /// </summary>
-    public class AreaConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-            {
-                if (d > 999999)
-                    return (d * 1E-6).ToString("N0", culture); // km^2
-                return d.ToString("N0", culture);
-            }
-
-            return string.Empty;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    /// <summary>
-    ///     If value is higher than 999999 m2, than show square km2
-    /// </summary>
-    public class AreaConverterUnit : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is double d)
-                if (d > 999999)
-                    return "km²"; // km^2
-            return "m²";
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
     }
 
     public static class NumericExtensions
