+++ DistrictEnergy/Views/ResultViews/Summary.xaml.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+using DistrictEnergy.ViewModels;
+
+namespace DistrictEnergy.Views.ResultViews
+{
+    /// <summary>
+    /// Interaction logic for Summary.xaml
+    /// </summary>
+    public partial class Summary : UserControl
+    {
+        public Summary()
+        {
+            InitializeComponent();
+
+            DataContext = new SummaryViewModel();
+        }
+    }
+}
